move helper to HTTP utils module
